Исправил ворнинги и утечку памяти

+++ TestHeader.h
@@ -1,5 +1,6 @@
 #ifndef TEST_HEADER_H_
 #define TEST_HEADER_H_
+
 #include "SquareSolver.h"
 
 struct test
@@ -9,8 +10,10 @@ struct test
     double x2_test;
     int num_test;
     };
-void unit_test(double* ptr_data, int length);
+
+void unit_test(double* ptr_data, size_t length);
 int admin_function();
-//struct koeficients scan_struct_test_function(double scan_a, double scan_b, double scan_c);
 void init_test(struct test library_unit_test);
+//struct koeficients scan_struct_test_function(double scan_a, double scan_b, double scan_c);
+
 #endif
